+++ gwcelery/tests/test_tasks_external_triggers.py
@@ -28,7 +28,8 @@ def test_handle_create_grb_event(mock_create_event, mock_get_event,
     mock_create_event.assert_called_once_with(filecontents=text,
                                               search='GRB',
                                               pipeline='Fermi',
-                                              group='External')
+                                              group='External',
+                                              labels=None)
     calls = [
         call(
             '"dqrjson"', 'gwcelerydetcharcheckvectors-E1.json', 'E1',
@@ -99,12 +100,47 @@ def test_handle_create_subthreshold_grb_event(mock_get_upload_ext_skymap,
     mock_create_event.assert_called_once_with(filecontents=text,
                                               search='SubGRB',
                                               pipeline='Fermi',
-                                              group='External')
+                                              group='External',
+                                              labels=None)
     mock_check_vectors.assert_called_once()
-    mock_get_upload_ext_skymap.assert_called_once()
 
 
 @patch('gwcelery.tasks.external_skymaps.get_upload_external_skymap')
+@patch('gwcelery.tasks.gracedb.get_events', return_value=[])
+@patch('gwcelery.tasks.gracedb.get_event', return_value={
+    'graceid': 'E1', 'gpstime': 1, 'instruments': '', 'pipeline': 'Fermi',
+    'extra_attributes': {'GRB': {'trigger_duration': 1, 'trigger_id': 123,
+                                 'ra': 0., 'dec': 0., 'error_radius': 10.}},
+    'links': {'self': 'https://gracedb.ligo.org/events/E356793/'}})
+@patch('gwcelery.tasks.gracedb.create_event')
+@patch('gwcelery.tasks.detchar.check_vectors')
+def test_handle_noise_fermi_event(mock_check_vectors,
+                                  mock_create_event,
+                                  mock_get_event,
+                                  mock_get_events,
+                                  mock_get_upload_external_skymap):
+    text = resource_string(__name__, 'data/fermi_noise_gcn.xml')
+    external_triggers.handle_grb_gcn(payload=text)
+    mock_get_events.assert_called_once_with(query=(
+                                            'group: External pipeline: '
+                                            'Fermi grbevent.trigger_id '
+                                            '= "598032876"'))
+    # Note that this is the exact ID in the .xml file
+    mock_create_event.assert_called_once_with(filecontents=text,
+                                              search='GRB',
+                                              pipeline='Fermi',
+                                              group='External',
+                                              labels=['NOT_GRB'])
+    mock_check_vectors.assert_called_once()
+    mock_get_upload_external_skymap.assert_called_once()
+
+
+@pytest.mark.parametrize('filename',
+                         ['data/fermi_grb_gcn.xml',
+                          'data/fermi_noise_gcn.xml'])
+@patch('gwcelery.tasks.external_skymaps.get_upload_external_skymap')
+@patch('gwcelery.tasks.gracedb.create_label')
+@patch('gwcelery.tasks.gracedb.remove_label')
 @patch('gwcelery.tasks.gracedb.replace_event')
 @patch('gwcelery.tasks.gracedb.get_events', return_value=[{
     'graceid': 'E1', 'gpstime': 1, 'instruments': '', 'pipeline': 'Fermi',
@@ -117,11 +153,16 @@ def test_handle_create_subthreshold_grb_event(mock_get_upload_ext_skymap,
                                  'ra': 10., 'dec': 0., 'error_radius': 10.}},
     'links': {'self': 'https://gracedb.ligo.org/events/E356793/'}})
 def test_handle_replace_grb_event(mock_get_event, mock_get_events,
-                                  mock_replace_event,
-                                  mock_get_upload_external_skymap):
-    text = resource_string(__name__, 'data/fermi_grb_gcn.xml')
+                                  mock_replace_event, mock_remove_label,
+                                  mock_create_label,
+                                  mock_get_upload_external_skymap, filename):
+    text = resource_string(__name__, filename)
     external_triggers.handle_grb_gcn(payload=text)
     mock_replace_event.assert_called_once_with('E1', text)
+    if 'grb' in filename:
+        mock_remove_label.assert_called_once_with('NOT_GRB', 'E1')
+    elif 'noise' in filename:
+        mock_create_label.assert_called_once_with('NOT_GRB', 'E1')
 
 
 @patch('gwcelery.tasks.gracedb.get_group', return_value='CBC')